requirements.txt
pyirsdk>=1.3.0
openai>=1.40.0
anthropic>=0.34.0
google-generativeai>=0.8.0
requests>=2.31.0
elevenlabs>=1.0.0
edge-tts>=6.1.0
pygame-ce>=2.5.0
